+++ sys/platform/vkernel64/platform/pmap.c
@@ -2460,7 +2460,7 @@ pmap_enter_quick(pmap_t pmap, vm_offset_t va, vm_page_t m)
  * to be used for panic dumps.
  */
 void *
-pmap_kenter_temporary(vm_paddr_t pa, int i)
+pmap_kenter_temporary(vm_paddr_t pa, long i)
 {
 	pmap_kenter(crashdumpmap + (i * PAGE_SIZE), pa);
 	return ((void *)crashdumpmap);
